refac: 윈도우 사이즈에 따른 화면 조정 공통 함수 추출
